# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cross App Backend",
    "description": "API para gestao e gamificacao de CrossFit",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desenvolvimento local"
    }
  ],
...
  "paths": {
...
    "/wods": {
      "post": {
        "tags": [
          "Wods"
        ],
        "summary": "Cadastra o WOD do box",
        "description": "Recebe o JSON completo do treino e persiste na colecao wods vinculado ao boxId do ADMIN autenticado.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XBoxIdHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWodDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "WOD criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateWodResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados invalidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, invalido ou expirado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Perfil sem permissao para criar WOD",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Ja existe WOD para a data informada neste box",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/wods/today": {
      "get": {
        "tags": [
          "Wods"
        ],
        "summary": "Busca o WOD de hoje do box autenticado",
        "description": "Retorna o treino da data atual associado ao boxId presente no token JWT, para exibicao no app do aluno.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XBoxIdHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "WOD de hoje retornado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/Wod"
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, invalido ou expirado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "XBoxIdHeader": {
        "name": "x-box-id",
        "in": "header",
        "required": true,
        "description": "ID do box selecionado",
        "schema": {
          "type": "string",
          "example": "67ea76a5ac5d89c8bb9d2111"
        }
      }
    },
    "schemas": {
      "ObjectId": {
        "type": "string",
        "pattern": "^[a-fA-F0-9]{24}$",
        "example": "67ea76a5ac5d89c8bb9d2111"
      },
...
      "WodBlockType": {
        "type": "string",
        "enum": [
          "WARMUP",
          "SKILL",
          "WOD"
        ]
      },
...
      "CreateWodBlockDto": {
        "type": "object",
        "required": [
          "type",
          "title",
          "content"
        ],
        "properties": {
          "type": {
            "$ref": "#/components/schemas/WodBlockType"
          },
          "title": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120,
            "example": "Warm-up geral"
          },
          "content": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2000,
            "example": "3 rounds: 20 air squats, 10 inchworms, 200m run"
          }
        }
      },
      "CreateWodDto": {
        "type": "object",
        "required": [
          "date",
          "title",
          "blocks"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "example": "2026-03-31"
          },
          "title": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120,
            "example": "WOD de Terca"
          },
          "blocks": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/CreateWodBlockDto"
            }
          }
        }
      },
...
      "WodBlock": {
        "type": "object",
        "required": [
          "type",
          "title",
          "content"
        ],
        "properties": {
          "type": {
            "$ref": "#/components/schemas/WodBlockType"
          },
          "title": {
            "type": "string",
            "example": "Warm-up geral"
          },
          "content": {
            "type": "string",
            "example": "3 rounds: 20 air squats, 10 inchworms, 200m run"
          }
        }
      },
      "Wod": {
        "type": "object",
        "required": [
          "_id",
          "boxId",
          "date",
          "title",
          "blocks",
          "createdAt"
        ],
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "boxId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "title": {
            "type": "string",
            "example": "WOD de Terca"
          },
          "blocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WodBlock"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateWodResponse": {
        "type": "object",
        "required": [
          "wodId",
          "boxId"
        ],
        "properties": {
          "wodId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "boxId": {
            "$ref": "#/components/schemas/ObjectId"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "statusCode": {
            "type": "number",
            "example": 400
          }
        }
      },
...
    }
  }
}